requirements.txt
https://download.pytorch.org/whl/cu121/torch-2.2.1%2Bcu121-cp310-cp310-win_amd64.whl
https://download.pytorch.org/whl/cu121/torchvision-0.17.1%2Bcu121-cp310-cp310-win_amd64.whl
https://download.pytorch.org/whl/cu121/torchaudio-2.2.1%2Bcu121-cp310-cp310-win_amd64.whl
git+https://github.com/m-bain/whisperx.git
fastapi
uvicorn
python-multipart
requests
python-dotenv